+++ ml-agents/mlagents/trainers/torch/utils.py
@@ -8,6 +8,7 @@ from mlagents.trainers.torch.encoders import (
     ResNetVisualEncoder,
     NatureVisualEncoder,
     SmallVisualEncoder,
+    FullyConnectedVisualEncoder,
     VectorInput,
 )
 from mlagents.trainers.settings import EncoderType, ScheduleType
@@ -20,6 +21,7 @@ class ModelUtils:
     # Minimum supported side for each encoder type. If refactoring an encoder, please
     # adjust these also.
     MIN_RESOLUTION_FOR_ENCODER = {
+        EncoderType.FULLY_CONNECTED: 1,
         EncoderType.MATCH3: 5,
         EncoderType.SIMPLE: 20,
         EncoderType.NATURE_CNN: 36,
@@ -123,6 +125,7 @@ class ModelUtils:
             EncoderType.NATURE_CNN: NatureVisualEncoder,
             EncoderType.RESNET: ResNetVisualEncoder,
             EncoderType.MATCH3: SmallVisualEncoder,
+            EncoderType.FULLY_CONNECTED: FullyConnectedVisualEncoder,
         }
         return ENCODER_FUNCTION_BY_TYPE.get(encoder_type)